feat(model): Added cloud storage integration to retrieve documents
Added Google Drive and Dropbox integration that allows users to select PDF files stored on their cloud storage.
This improves usability and user interaction as shared files can be used as documents to query upon.
Moreover this change extends the features and functionality of the existing application providing a new source of data and/or documents through cloud storage.