docs: Complete the version 2.0 implementation
This change only modifies checking of the features list that had to be
implemented. This change concludes the implementation of Version 2.0,
which was planned and scheduled. The documentation of the current
version is still pending, along with the implementation of the approved architecturral structure modfications.

Next Task: Documentation of Code